forge fmt

+++ test/util/mock/MockFillContract.sol
@@ -9,7 +9,7 @@ contract MockFillContract is IReactorCallback {
     /// @notice assume that we already have all output tokens
     function reactorCallback(ResolvedOrder[] memory resolvedOrders, bytes memory) external {
         for (uint256 i = 0; i < resolvedOrders.length; i++) {
-            for(uint j = 0; j < resolvedOrders[i].outputs.length; j++) {
+            for (uint256 j = 0; j < resolvedOrders[i].outputs.length; j++) {
                 ERC20 token = ERC20(resolvedOrders[i].outputs[j].token);
                 token.approve(msg.sender, type(uint256).max);
             }
